ocr-server/requirements.txt
easyocr>=1.7.2
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pillow>=11.0.0
